requirements.txt
discord.py
python-dotenv
mysqlclient
SQLAlchemy
requests
